Refactored BLT commands to consolidate Acquia API reliant commands
