+++ blt/src/Blt/Plugin/Commands/HsCommands.php
@@ -2,24 +2,13 @@
 
 namespace Humsci\Blt\Plugin\Commands;
 
+use Acquia\Blt\Robo\BltTasks;
 use Drupal\Core\Serialization\Yaml;
-use Symfony\Component\Console\Question\Question;
-
-if (!trait_exists('Sws\BltSws\Blt\Plugin\Commands\SwsCommandTrait')) {
-  return;
-}
 
 /**
  * Various BLT commands for H&S stack.
  */
-class HsCommands extends HsAcquiaApiCommands {
-
-  /**
-   * List of failed databases from the sync command.
-   *
-   * @var array
-   */
-  protected $failedDatabases = [];
+class HsCommands extends BltTasks {
 
   /**
    * Set up local blt settings and necessary files.
@@ -70,7 +59,7 @@ class HsCommands extends HsAcquiaApiCommands {
   }
 
   /**
-   * Get encryption keys from acquia.
+   * Send encryption keys to acquia.
    *
    * @param string $env
    *   Acquia environment to send the keys.
@@ -84,34 +73,6 @@ class HsCommands extends HsAcquiaApiCommands {
       $this->taskDrush()
         ->drush("rsync @self:../keys/ @default.$env:$key_dir")
         ->run();
-    }
-  }
-
-  /**
-   * Create a new database on Acquia environment.
-   *
-   * @command humsci:create-database
-   */
-  public function createDatabase() {
-    $database = $this->getMachineName('What is the name of the database? This ideally will match the site directory name. No special characters please.');
-    $this->connectAcquiaApi();
-    $this->say('<info>' . $this->acquiaDatabases->create($this->appId, $database)->message . '</info>');
-  }
-
-  /**
-   * Add a domain to Acquia environment.
-   *
-   * @param string $environment
-   *   Environment: dev, test, or prod.
-   * @param string $domains
-   *   Comma separated new domain to add.
-   *
-   * @command humsci:add-domain
-   */
-  public function humsciAddDomain($environment, $domains) {
-    $this->connectAcquiaApi();
-    foreach (explode(',', $domains) as $domain) {
-      $this->say($this->acquiaDomains->create($this->getEnvironmentUuid($environment), $domain)->message);
     }
   }
 
@@ -243,54 +204,6 @@ class HsCommands extends HsAcquiaApiCommands {
   }
 
   /**
-   * Sync the staging sites databases with that from production.
-   *
-   * @command humsci:sync-stage
-   *
-   * @options exclude Comma separated list of database names to skip.
-   */
-  public function syncStaging(array $options = ['exclude' => NULL]) {
-    $task_started = time() - (60 * 60 * 24);
-    $this->connectAcquiaApi();
-
-    $sites = $this->getSitesToSync($task_started, $options);
-    if (!$this->confirm(sprintf('Are you sure you wish to stage the following sites: <comment>%s</comment>', implode(', ', $sites)))) {
-      return;
-    }
-    $count = count($sites);
-    $copy_sites = array_splice($sites, 0, 4);
-
-    foreach ($copy_sites as $site) {
-      $this->say("Copying $site database to staging.");
-      $this->acquiaDatabases->copy($this->getEnvironmentUuid('prod'), $site, $this->getEnvironmentUuid('test'));
-    }
-
-    while (!empty($sites)) {
-      echo '.';
-      sleep(10);
-      $finished_databases = $this->getCompletedDatabaseCopies($task_started);
-
-      if ($finished = array_intersect($copy_sites, $finished_databases)) {
-        echo PHP_EOL;
-        foreach ($finished as $copied_db) {
-          $db_position = array_search($copied_db, $copy_sites);
-          $new_site = array_splice($sites, 0, 1);
-          $new_site = reset($new_site);
-          $copy_sites[$db_position] = $new_site;
-          $this->say("Copying $new_site database to staging.");
-          $this->connectAcquiaApi();
-          $this->say($this->acquiaDatabases->copy($this->getEnvironmentUuid('prod'), $new_site, $this->getEnvironmentUuid('test'))->message);
-        }
-      }
-    }
-    $this->yell("$count database have been copied to staging.");
-
-    if (array_unique($this->failedDatabases)) {
-      $this->yell('Databases failed: ' . implode(', ', array_unique($this->failedDatabases)), 40, 'red');
-    }
-  }
-
-  /**
    * Get an overall list of database names to sync.
    *
    * @param int $task_started
@@ -317,76 +230,6 @@ class HsCommands extends HsAcquiaApiCommands {
       $sites = array_diff($sites, $exclude);
     }
     return array_diff($sites, $finished_databases);
-  }
-
-  /**
-   * Get a list of all databases that have finished copying after a time.
-   *
-   * @param int $time_comparison
-   *   Time to compare the completed task.
-   *
-   * @return array
-   *   Array of database names.
-   */
-  protected function getCompletedDatabaseCopies($time_comparison) {
-    $databases = [];
-    $this->connectAcquiaApi();
-    /** @var \AcquiaCloudApi\Response\NotificationResponse $notification */
-    foreach ($this->acquiaNotifications->getAll($this->appId) as $notification) {
-      if (
-        isset($notification->event) &&
-        $notification->event == 'DatabaseCopied' &&
-        strtotime($notification->created_at) >= $time_comparison
-      ) {
-        if ($notification->status == 'completed') {
-          $databases = array_merge($databases, $notification->context->database->names);
-        }
-        else {
-          $this->failedDatabases = array_merge($this->failedDatabases, $notification->context->database->names);
-        }
-      }
-    }
-    return array_values(array_unique($databases));
-  }
-
-  /**
-   * {@inheritDoc}
-   */
-  protected function connectAcquiaApi() {
-    if (!$this->acquiaApplications) {
-      parent::connectAcquiaApi();
-      return;
-    }
-    try {
-      $this->acquiaApplications->getAll();
-    }
-    catch (\Exception $e) {
-      parent::connectAcquiaApi();
-    }
-  }
-
-  /**
-   * Ask the user for a new stanford url and validate the entry.
-   *
-   * @param string $message
-   *   Prompt for the user.
-   *
-   * @return string
-   *   User entered value.
-   */
-  protected function getMachineName($message) {
-    $question = new Question($this->formatQuestion($message));
-    $question->setValidator(function ($answer) {
-      $modified_answer = strtolower($answer);
-      $modified_answer = preg_replace("/[^a-z0-9_]/", '_', $modified_answer);
-      if ($modified_answer != $answer) {
-        throw new \RuntimeException(
-          'Only lower case alphanumeric characters with underscores are allowed.'
-        );
-      }
-      return $answer;
-    });
-    return $this->doAsk($question);
   }
 
 }
